Update to reference Open Shift Variables
There appears to be a great deal of user issues with actually accessing the open shift MySQL database.  This should help people get up and started connecting to the database without the headache of tracing down this issue.  Users are attempting to hard-code these values and struggling.  Hopefully this saves people a lot of trouble.